upipe-hbrmt: move C packing/encoding function to separate file
