+++ lib/upipe-hbrmt/Makefile.am
@@ -4,6 +4,8 @@ libupipe_hbrmt_la_SOURCES = upipe_pack10bit.c \
     upipe_unpack10bit.c \
     sdidec.c \
     sdidec.h \
+    sdienc.c \
+    sdienc.h \
     $(NULL)
 
 libupipe_hbrmt_la_CPPFLAGS = -I$(top_builddir)/include -I$(top_srcdir)/include
@@ -13,8 +15,7 @@ libupipe_hbrmt_la_LDFLAGS = -no-undefined
 
 if HAVE_X86ASM
 libupipe_hbrmt_la_SOURCES += sdidec.asm \
-    sdienc.asm \
-    sdienc.h
+    sdienc.asm
 endif
 
 pkgconfigdir = $(libdir)/pkgconfig
