+++ lib/upipe-hbrmt/sdienc.h
@@ -1,3 +1,4 @@
+void upipe_sdi_pack_c(uint8_t *dst, const uint8_t *y, int64_t size);
 void upipe_sdi_pack_10_ssse3(uint8_t *dst, const uint8_t *y, int64_t size);
 void upipe_sdi_pack_10_avx  (uint8_t *dst, const uint8_t *y, int64_t size);
 void upipe_sdi_pack_10_avx2 (uint8_t *dst, const uint8_t *y, int64_t size);
